+++ src/Traits/Helpers/SearchHelpers.php
@@ -88,4 +88,15 @@ trait SearchHelpers
 
         return '';
     }
+
+    public function getSearchPlaceholder(): string
+    {
+        return $this->hasSearchPlaceholder() ? $this->searchPlaceholder : __('Search');
+    }
+
+    public function hasSearchPlaceholder(): bool
+    {
+        return $this->searchPlaceholder !== null;
+    }
+
 }