+++ src/Traits/WithSearch.php
@@ -22,6 +22,8 @@ trait WithSearch
     public ?bool $searchFilterDefer = null;
 
     public ?bool $searchFilterLazy = null;
+    
+    public ?string $searchPlaceholder = null;
 
     // TODO
     public function applySearch(): Builder