+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to `laravel-livewire-tables` will be documented in this file
 
 ## [Unreleased]
+- Enhancements
+    - Add customisable Search Placeholder - setSearchPlaceholder()
 
 ## [2.15.0] - 2023-07-15